feat: adjust ui register screen and rename file screen reset password
